+++ sei-ibc-go/modules/apps/transfer/types/packet.go
@@ -23,7 +23,7 @@ var (
 
 // NewFungibleTokenPacketData contructs a new FungibleTokenPacketData instance
 func NewFungibleTokenPacketData(
-	denom string, amount uint64,
+	denom string, amount string,
 	sender, receiver string,
 ) FungibleTokenPacketData {
 	return FungibleTokenPacketData{
@@ -38,8 +38,12 @@ func NewFungibleTokenPacketData(
 // NOTE: The addresses formats are not validated as the sender and recipient can have different
 // formats defined by their corresponding chains that are not known to IBC.
 func (ftpd FungibleTokenPacketData) ValidateBasic() error {
-	if ftpd.Amount == 0 {
-		return sdkerrors.Wrap(ErrInvalidAmount, "amount cannot be 0")
+	amount, ok := sdk.NewIntFromString(ftpd.Amount)
+	if !ok {
+		return sdkerrors.Wrapf(ErrInvalidAmount, "unable to parse transfer amount (%s) into sdk.Int", ftpd.Amount)
+	}
+	if !amount.IsPositive() {
+		return sdkerrors.Wrapf(ErrInvalidAmount, "amount must be strictly positive: got %d", amount)
 	}
 	if strings.TrimSpace(ftpd.Sender) == "" {
 		return sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, "sender address cannot be blank")