+++ sei-ibc-go/modules/apps/transfer/types/packet_test.go
@@ -7,8 +7,10 @@ import (
 )
 
 const (
-	denom  = "transfer/gaiachannel/atom"
-	amount = uint64(100)
+	denom              = "transfer/gaiachannel/atom"
+	amount             = "100"
+	largeAmount        = "18446744073709551616"                                                           // one greater than largest uint64 (^uint64(0))
+	invalidLargeAmount = "115792089237316195423570985008687907853269984665640564039457584007913129639936" // 2^256
 )
 
 // TestFungibleTokenPacketDataValidateBasic tests ValidateBasic for FungibleTokenPacketData
@@ -19,8 +21,12 @@ func TestFungibleTokenPacketDataValidateBasic(t *testing.T) {
 		expPass    bool
 	}{
 		{"valid packet", NewFungibleTokenPacketData(denom, amount, addr1, addr2), true},
+		{"valid packet with large amount", NewFungibleTokenPacketData(denom, largeAmount, addr1, addr2), true},
 		{"invalid denom", NewFungibleTokenPacketData("", amount, addr1, addr2), false},
-		{"invalid amount", NewFungibleTokenPacketData(denom, 0, addr1, addr2), false},
+		{"invalid empty amount", NewFungibleTokenPacketData(denom, "", addr1, addr2), false},
+		{"invalid zero amount", NewFungibleTokenPacketData(denom, "0", addr1, addr2), false},
+		{"invalid negative amount", NewFungibleTokenPacketData(denom, "-1", addr1, addr2), false},
+		{"invalid large amount", NewFungibleTokenPacketData(denom, invalidLargeAmount, addr1, addr2), false},
 		{"missing sender address", NewFungibleTokenPacketData(denom, amount, emptyAddr, addr2), false},
 		{"missing recipient address", NewFungibleTokenPacketData(denom, amount, addr1, emptyAddr), false},
 	}